Add django-less miner client in lib
